requirements.txt
streamlit==1.37.1
torch==2.4.1
torchvision==0.19.1
Pillow==10.4.0
pytest==8.3.2
